rework PFFile
